+++ src/papi/papi_wrapper.sh
@@ -1,14 +1,17 @@
+#!/usr/bin/env sh
 # ------------------------------------------------------------------------------
 # This wrapper contains functions for interacting with the PAPI energy
 # profiling options.
 # ------------------------------------------------------------------------------
 
 # Get the directory where this file is located to load dependencies.
-PAPIDIR="$BASEDIR/papi"
+PAPIDIR="$SRCDIR/papi"
 . "$PAPIDIR/../utils/general_utils.sh"
 . "$PAPIDIR/../utils/print_utils.sh"
+. "$PAPIDIR/../utils/array_utils.sh"
 
 PAPI_PROFILER="$PAPIDIR/papi_profiler.o"
+
 
 # Returns 0 if papi is available, 1 otherwise.
 papi_available() {
@@ -25,10 +28,16 @@ papi_available() {
 }
 
 _compile_papi_profiler() {
-    # Check if the binary exists, if so remove.
+    # Check if the binary exists.
     if [ -f "$PAPI_PROFILER" ]; then
-        rm "$PAPI_PROFILER"
+        # If exists, check for executable rights and rebuild if not.
+        if [ -x "$PAPI_PROFILER" ]; then
+            return
+        else
+            rm "$PAPI_PROFILER"
+        fi
     fi
+
 
     # Compile the code.
     if ! cc "$PAPIDIR/papi_profiler.c" "$PAPIDIR/papi_component.c" "$PAPIDIR/papi_event.c" -o "$PAPI_PROFILER" -lpapi -Wall; then
@@ -43,18 +52,23 @@ _compile_papi_profiler() {
     fi
 }
 
-# Convert a unit string into a floating-point scaling factor to Joules
-_parse_papi_unit_to_joules() {
-    local unit="$1"
+# Convert a unit-string into a floating-point scalar for converting in Joule.
+_parse_papi_unit_to_joule_scalar() {
+    unit="$1"
 
-    # Parse scientific notation like "2^-32 Joules"
-    if [[ "$unit" =~ ^([0-9.]+)\^(-?[0-9]+)$ ]]; then
-        local base="${BASH_REMATCH[1]}"
-        local exponent="${BASH_REMATCH[2]}"
-        # local exponent=$(echo "$unit" | sed -E "s/^${base}\^\((-?[0-9]+)\).*$/\1/")
-        printf "scale=20; %s^(%s)\n" "$base" "$exponent" | bc -l | sed 's/^\./0./'
+    case $unit in
+    *"^"*)
+        splitted=$(split_scientific_notation "$unit")
+        base=${splitted%%$'\n'*}
+        exponent=${splitted#*$'\n'}
+
+        # Compute the value using bc
+        printf "scale=20; %s^(%s)\n" "$base" "$exponent" |
+            bc -l |
+            sed 's/^\./0./'
         return
-    fi
+        ;;
+    esac
 
     # Parse SI prefixes.
     case "$unit" in
@@ -66,24 +80,85 @@ _parse_papi_unit_to_joules() {
     kJ)         printf "1000\n" ;;
     MJ)         printf "1000000\n" ;;
     GJ)         printf "1000000000\n" ;;
-    *)          printf "Unrecognized unit: %s\n" "$unit" >&2 ;;
+    *)          printf "Unrecognized SI prefix unit: %s\n" "$unit" >&2 ;;
     esac
 }
 
 # Print papi event information depending on the mode given.
 _print_papi_event() {
-    local mode="$1"
-    local event="$2"
-    local unit="$3"
+    mode="$1"
+    event="$2"
+    unit="$3"
 
     # Print found values.
     if [ "$mode" = "events" ]; then
         printf "%s\n" "$event"
     elif [ "$mode" = "units" ]; then
-        ecprintf "%s\n"ho "$unit"
+        printf "%s\n" "$unit"
     else
         printf "%s : %s\n" "$event" "$unit"
     fi
+}
+
+# Detect the start of a PAPI event block with energy counter information.
+_detect_papi_start_event_block() {
+    # Reject UNITS events.
+    case $1 in
+        '|'[[:space:]]*cray_rapl:::UNITS* | \
+        '|'[[:space:]]*cray_pm:::UNITS* )
+            return 1
+            ;;
+    esac
+
+    # Accept RAPL/CRAY_PM energy events.
+    case $1 in
+        '|'[[:space:]]*rapl::* | \
+        '|'[[:space:]]*cray_rapl::* | \
+        '|'[[:space:]]*cray_pm:::PM_ENERGY:* )
+            return 0
+            ;;
+    esac
+
+    # Return false in any other case.
+    return 1
+}
+
+# Detect a modifier line inside a PAPI event block, like :cpu=0 or :freq=2000.
+_detect_papi_modifier_line() {
+    case $1 in
+        '|'[[:space:]]*:*=* )
+            return 0
+            ;;
+    esac
+
+    return 1
+}
+
+# Detect a unit line inside a PAPI event block in the style of MN5 and LUMI.
+_detect_unit_line() {
+    case $1 in
+        *[Uu]nit* )
+            # Must contain at least one of these separators:
+            case $1 in
+                *:[[:space:]]*|*is[[:space:]]* ) return 0 ;;
+            esac
+            ;;
+    esac
+
+    return 1
+}
+
+# Detect the end of a PAPI event block.
+_detect_event_block_end() {
+    case $1 in
+        # One or more '-' or '=' only
+        [-=]* )
+            # But it must not be empty
+            [ -n "$1" ] && return 0
+            ;;
+    esac
+
+    return 1
 }
 
 # Parse the output of papi_native_avail for energy related events and units.
@@ -92,67 +167,102 @@ _parse_papi_native_avail() {
     #   - "events": Only print the event names.
     #   - "units":  Only print the units.
     #   - "both":   Print both names and units separated by " : ".
-    local mode="$1"
+    mode="$1"
 
     # Get events from user.
-    local events="$2"
+    events="$2"
 
-    # Locals for parsing each event.
-    local in_event=0
-    local event_name=""
-    local unit=""
-    local modifiers=()
-    local mod=""
+    # Locals for parsing each event, where unit will contain Joule scalars.
+    in_event=0
+    event_name=""
+    unit=""
 
     # Loop over all lines of the output.
-    echo "$events" | while IFS= read -r line; do
-        # Detect the start of a rapl or cray_pm event block.
-        if [[ "$line" =~ ^\|[[:space:]]*(rapl::|cray_rapl::|cray_pm:::PM_ENERGY:)[^[:space:]]+ ]]; then
-            # Exclude any UNITS events.
-            if [[ ! "$line" =~ ^\|[[:space:]]*(cray_rapl:::UNITS|cray_pm:::UNITS) ]]; then
-                # Extract the event name.
-                event_name=$(printf "%s\n" "$line" | awk -F'|' '{gsub(/^[ \t]+|[ \t]+$/, "", $2); print $2}')
+    printf '%s\n' "$events" | while IFS= read -r line; do
+        # Detect the start of a rapl or cray_pm event block, which is not a
+        # UNITS event.
+        if _detect_papi_start_event_block "$line"; then
+            # Extract the event name.
+            event_name=$(printf "%s\n" "$line" \
+                | awk -F'|' '{gsub(/^[ \t]+|[ \t]+$/, "", $2); print $2}')
 
-                # Reset values for parsing this event.
-                in_event=1
-                modifiers=()
-                unit=""
-                continue
-            fi
+            # Reset values for parsing this event.
+            in_event=1
+            modifiers=""
+            unit=""
+            continue
         fi
 
         # Parse consecutive lines after the event name line.
-        if [[ $in_event -eq 1 ]]; then
+        if [ $in_event -eq 1 ]; then
             # Extract modifiers, like :cpu=*.
-            if [[ "$line" =~ ^\|[[:space:]]*(:[a-zA-Z0-9_]+=[^[:space:]]+) ]]; then
-                mod="${BASH_REMATCH[1]}"
-                case "$mod" in
+            if _detect_papi_modifier_line "$line"; then
+                # Extract the modifier name, e.g. "cpu"
+                mod=$(printf '%s\n' "$line" \
+                    | sed -n 's/^|[[:space:]]*\(:[^[:space:]]*\).*$/\1/p')
+
+                # If extraction failed, skip
+                [ -z "$mod" ] && continue
+
                 # Skip known modifiers that do not make sense to alter.
-                :period=* | :freq=* | :excl=* | :pinned=*) ;;
-                *) modifiers+=("$mod") ;;
+                case $mod in
+                    :period=* | :freq=* | :excl=* | :pinned=* )
+                        ;;
+                    *)
+                        # Store parsed modifiers in an array.
+                        modifiers=$(array_push "$modifiers" "$mod")
+                        ;;
                 esac
             fi
 
             # Extract units in MN5 and LUMI format.
-            if [[ "$line" =~ Units?[[:space:]]*(:|is)[[:space:]]*([0-9]+|\-?[0-9]+|2\^[\-0-9]+|[a-zA-Z]+) ]]; then
-                unit=$(_parse_papi_unit_to_joules "${BASH_REMATCH[2]}")
+            if _detect_unit_line "$line"; then
+                # Extract unit token and convert to scalar.
+                case $line in
+                    # Parse MN5 format.
+                    *'Unit is '*)
+                        unit_token=$(printf '%s\n' "$line" |
+                            sed 's/.*Unit is[[:space:]]*//; s/[[:space:]]*|.*//')
+                        scalar=${unit_token%% *}
+                        unit=${unit_token#* }
+
+                        # Print warning if unit is not Joules.
+                        if [ "$unit" == "Joules" ]; then
+                            unit=$(_parse_papi_unit_to_joule_scalar "$scalar")
+                        else
+                            verbose_echo print_warning "Unit line did not have Joules unit."
+                        fi
+                        ;;
+                    # Parse LUMI format.
+                    *'Units:'*)
+                        unit=$(printf '%s\n' "$line" |
+                            sed 's/.*Units:[[:space:]]*//; s/[[:space:]]*|.*//')
+                        unit=$(_parse_papi_unit_to_joule_scalar "$unit")
+                        ;;
+                    *)
+                        verbose_echo print_warning "Could not parse unit line.."
+                        ;;
+                esac
             fi
 
             # Detect the end of the event block.
-            if [[ "$line" =~ ^[-=]+$ ]]; then
+            if _detect_event_block_end "$line"; then
                 # Default units to 1 if none was detected.
                 [ -z "$unit" ] && unit=1
 
-                # Take combinations of the event_name and modifiers, if any exist.
-                if [ ${#modifiers[@]} -eq 0 ]; then
+                # If no modifiers were collected, provide event as-is.
+                if [ "$(array_len "$modifiers")" -eq 0 ]; then
                     _print_papi_event "$mode" "$event_name" "$unit"
                 else
-                    for mod in "${modifiers[@]}"; do
-                        _print_papi_event "$mode" "$event_name$mod" "$unit"
+                    # Otherwise, take combinations of the event and modifiers.
+                    while [ "$(array_len "$modifiers")" -ne 0 ]; do
+                        mod=$(array_get_last "$modifiers")
+                        modifiers=$(array_pop "$modifiers")
+                        _print_papi_event "$mode" "${event_name}${mod}" "$unit"
                     done
                 fi
 
-                # Reset event block detection.
+                # Reset event block detection state.
                 in_event=0
             fi
         fi
@@ -171,21 +281,26 @@ _get_papi_native_avail() {
     #   - "events": Only print the event names.
     #   - "units":  Only print the units.
     #   - "both":   Print both names and units separated by " : ".
-    local mode="$1"
+    mode="$1"
 
     # Fetch list of counters and units.
-    local components=("rapl" "cray_pm")
-    local events=""
+    components=""
+    components=$(array_push "$components" "rapl")
+    components=$(array_push "$components" "cray_pm")
+    events=""
 
-    for component in "${components[@]}"; do
+    while [ "$(array_len "$components")" -ne 0 ]; do
+        component=$(array_get_last "$components")
+        components=$(array_pop "$components")
         events=$(papi_native_avail -i "$component")
+
+        # Print the found counters and units to stdout.
         _parse_papi_native_avail "$mode" "$events"
     done
 }
 
 # Return the set of energy events supported by this system.
 papi_events() {
-    local events
     events=$(_get_papi_native_avail "events")
     if [ -z "$events" ]; then
         printf "NO EVENTS AVAILABLE\n"
@@ -203,7 +318,6 @@ papi_profile() {
     fi
 
     # Get events and units.
-    local events units
     events=$(_get_papi_native_avail "events")
     units=$(_get_papi_native_avail "units")
